feat: added economic calendar configrations to github and added logic for updating the login links
